progress on rl ai implementation + BUG
bug when playing numeral cards, e.g. playing P1A on first move is right, but can play this as P1A1, P1A2, etc. also, as though the numeral P1 is a face card. causes issues for rl as it is randomly selecting one of these erroneous P options. need to work on that.